+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.2.2] - 2024-04-10
+
 ### Added
 
 - Automatic resource injection of [`@popperjs/core`](https://www.npmjs.com/package/@popperjs/core) for PatternFly Java GWT
@@ -273,7 +275,8 @@ Stay tuned for more to come...
 - for dependency upgrades
 -->
 
-[Unreleased]: https://github.com/patternfly-java/patternfly-java/compare/v0.2.1...HEAD
+[Unreleased]: https://github.com/patternfly-java/patternfly-java/compare/v0.2.2...HEAD
+[0.2.2]: https://github.com/patternfly-java/patternfly-java/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/patternfly-java/patternfly-java/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/patternfly-java/patternfly-java/compare/v0.1.5...v0.2.0
 [0.1.5]: https://github.com/patternfly-java/patternfly-java/compare/v0.1.4...v0.1.5